+++ src/main/java/com/greenfarm/greenfarmsystem/service/impl/OrganicInputTypeServiceImpl.java
@@ -1,9 +1,9 @@
 package com.greenfarm.greenfarmsystem.service.impl;
 
+import com.greenfarm.greenfarmsystem.exception.ExceptionNotFoundModel;
 import com.greenfarm.greenfarmsystem.model.OrganicInputTypeEntity;
 import com.greenfarm.greenfarmsystem.repository.OrganicInputTypeRepository;
 import com.greenfarm.greenfarmsystem.service.OrganicInputTypeService;
-import corp.sche.trmg.commons.exception.BusinessException;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -13,8 +13,6 @@ import javax.validation.Validation;
 import javax.validation.Validator;
 import javax.validation.ValidatorFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -25,9 +23,10 @@ public class OrganicInputTypeServiceImpl implements OrganicInputTypeService {
 
   @Override
   public OrganicInputTypeEntity findById(Long id) {
-    Optional<OrganicInputTypeEntity> optionalOrganicInputTypeEntity = organicInputTypeRepository.findById(id);
-    if(!optionalOrganicInputTypeEntity.isPresent()){
-      throw new BusinessException("", "Error de negocio", "Tipo de insumo organico no encontrado!: ");
+    Optional<OrganicInputTypeEntity> optionalOrganicInputTypeEntity = organicInputTypeRepository
+        .findById(id);
+    if (!optionalOrganicInputTypeEntity.isPresent()) {
+      throw new ExceptionNotFoundModel("TIPO DE INSUMO ORGANICO NO ENCONTRADO!, ID: " + id);
     }
     return optionalOrganicInputTypeEntity.get();
   }
@@ -42,8 +41,9 @@ public class OrganicInputTypeServiceImpl implements OrganicInputTypeService {
   public OrganicInputTypeEntity save(OrganicInputTypeEntity organicInputTypeEntity) {
     ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
     Validator validator = factory.getValidator();
-    Set<ConstraintViolation<OrganicInputTypeEntity>> violations = validator.validate(organicInputTypeEntity);
-    if(!violations.isEmpty()){
+    Set<ConstraintViolation<OrganicInputTypeEntity>> violations = validator
+        .validate(organicInputTypeEntity);
+    if (!violations.isEmpty()) {
       throw new ConstraintViolationException(violations);
     }
         /*
@@ -51,7 +51,8 @@ public class OrganicInputTypeServiceImpl implements OrganicInputTypeService {
             log.error(violation.getMessage());
         }
         * */
-    OrganicInputTypeEntity organicInputTypeEntity1 = organicInputTypeRepository.save(organicInputTypeEntity);
+    OrganicInputTypeEntity organicInputTypeEntity1 = organicInputTypeRepository
+        .save(organicInputTypeEntity);
     return organicInputTypeEntity1;
 
   }
@@ -62,7 +63,13 @@ public class OrganicInputTypeServiceImpl implements OrganicInputTypeService {
   }
 
   @Override
-  public void delete(Long id) {
+  public boolean delete(Long id) throws Exception {
+    Optional<OrganicInputTypeEntity> optionalOrganicInputTypeEntity = organicInputTypeRepository
+        .findById(id);
+    if (!optionalOrganicInputTypeEntity.isPresent()) {
+      throw new Exception("ID NO ENCONTRADO: " + id);
+    }
     organicInputTypeRepository.deleteById(id);
+    return true;
   }
 }
